+++ src/spellbot/client.py
@@ -45,8 +45,15 @@ class SpellBot(AutoShardedBot):
         self.create_connection = create_connection
         self.channel_locks = ExpiringDict(max_len=100, max_age_seconds=3600)  # 1 hr
 
+        # add debug logging for discord.py state
+        discord_state_logger = logging.getLogger("discord.state")
+        discord_state_logger.setLevel(logging.DEBUG)
+
     async def on_ready(self) -> None:
         logger.info("client ready")
+
+    async def on_shard_ready(self, shard_id: int) -> None:
+        logger.info(f"shard {shard_id} ready")
 
     async def setup_hook(self) -> None:
         # Note: In tests we create the connection using fixtures.